Refactoring codice Implementazione pattern strategy e lettura json attraverso fileReader
Cons: non è più possibile fare parse delle stringe formattate come json (ma non è necessario in quesot tipo di applicazione)

Validazione con test

Merge conflitti con avanzamento branch developement
