+++ src/main/java/com/aadm/cardexchange/server/jsonparser/CardParseStrategy.java
@@ -1,6 +1,6 @@
 package com.aadm.cardexchange.server.jsonparser;
 
-import com.aadm.cardexchange.shared.CardDecorator;
+import com.aadm.cardexchange.shared.models.CardDecorator;
 import com.google.gson.JsonObject;
 
 public interface CardParseStrategy {
