+++ src/test/java/com/aadm/cardexchange/CardExchangeSuite.java
@@ -2,6 +2,7 @@ package com.aadm.cardexchange;
 
 import com.aadm.cardexchange.client.HomeActivityTest;
 import com.aadm.cardexchange.server.CardServiceTest;
+import com.aadm.cardexchange.server.JSONParserTest;
 import com.aadm.cardexchange.shared.CardTest;
 import com.aadm.cardexchange.shared.MagicCardDecoratorTest;
 import com.aadm.cardexchange.shared.PokemonCardDecoratorTest;
@@ -15,6 +16,7 @@ import org.junit.platform.suite.api.Suite;
         YuGiOhCardDecoratorTest.class,
         PokemonCardDecoratorTest.class,
         MagicCardDecoratorTest.class,
+        JSONParserTest.class,
         CardServiceTest.class,
         HomeActivityTest.class,
 })
